+++ src/components/fui-checkbox/fui-checkbox.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { prefix } from '../prefix';
 import FuiIconCheck12X12 from '../../icons/fui-icon-check-12x12';
 import classnames from 'classnames';
-import FuiIconIntermediateLine2x10 from '../../icons/fui-icon-intermediate-line-2x10';
+import FuiIconIndeterminateLine2x10 from '../../icons/fui-icon-indeterminate-line-2x10';
 
 const compPrefix = `${prefix}-checkbox`;
 
 export interface FuiCheckboxProps {
   checked?: boolean
-  intermediate?: boolean
+  indeterminate?: boolean
   disabled?: boolean
   onToggle: (isChecked: boolean) => void
   checkLabel?: string
@@ -22,7 +22,7 @@ export const FuiCheckbox = (props: FuiCheckboxProps) => {
     `${compPrefix}-wrapper`,
     props.disabled ? `${prefix}-disabled` : `${prefix}-interactable`,
     props.className,
-    (props.checked || props.intermediate) ? 'checked' : ''
+    (props.checked || props.indeterminate) ? 'checked' : ''
   );
 
   const onClick = (e: React.MouseEvent<HTMLDivElement>) => {
@@ -41,7 +41,7 @@ export const FuiCheckbox = (props: FuiCheckboxProps) => {
   return (
     <div aria-label={props.ariaLabel} tabIndex={0} onKeyDown={onSpace} className={classNames} onClick={onClick}>
       <div role='checkbox' className={compPrefix} >
-        {props.checked ? <FuiIconCheck12X12 /> : props.intermediate ? <FuiIconIntermediateLine2x10 /> : null}
+        {props.checked ? <FuiIconCheck12X12 /> : props.indeterminate ? <FuiIconIndeterminateLine2x10 /> : null}
       </div>
       {props.checkLabel && <label>{props.checkLabel}</label>}
     </div>