+++ src/icons/fui-icon-indeterminate-line-2x10.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { prefix } from '../components/prefix';
 
-export default function FuiIconIntermediateLine2x10 () {
+export default function FuiIconIndeterminateLine2x10 () {
   return (
-    <svg className={`${prefix}-icon-intermediate-line-2x10`} width="18" height="18" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg">
+    <svg className={`${prefix}-icon-indeterminate-line-2x10`} width="18" height="18" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg">
       <path d="M4 9C4 8.44772 4.44772 8 5 8H13C13.5523 8 14 8.44772 14 9C14 9.55228 13.5523 10 13 10H5C4.44772 10 4 9.55228 4 9Z" fill="currentcolor" />
     </svg>
   );