+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2024-03-10
+* [FuiCheckbox] - Fixed naming ([#8](https://github.com/functional-ui/functional-ui-kit/issues/8))
+* [FuiTextInput / FuiRadio / FuiStepper] - Improved adherence to WCAG ([#7](https://github.com/functional-ui/functional-ui-kit/issues/7))
+* [Docs] - Upgrade to Storybook@7.6.17
+
 ## [0.1.0] - 2024-02-07
 * [FuiPopover] - Removed `react-tiny-popover` and replace with `floating-ui`
 * [FuiOptionGroup] - New Component 🎉